Removed LLVM compiler warnings

+++ syck/SyckInput.m
@@ -6,7 +6,7 @@
 
 void cocoa_syck_error_handler( SyckParser *p, const char *msg )
 {
-	NSLog(@"syck error:%s position:(%d, %d)", msg, p->linect, p->cursor - p->lineptr);
+	NSLog(@"syck error:%s position:(%d, %d)", msg, p->linect, (int)(p->cursor - p->lineptr));
 }
 
 SYMID cocoa_syck_parse_handler(SyckParser *p, SyckNode *n)
